House Prediction Project (Data Science Project 2)
For this project, I am working on this guided by a tutor from DigitalSkola. The goal of this project is to determine the best machine learning model to predict a house price in Taiwan based on several factors. This is an end-to-end project done by following the guidance of CRISP-DM.